# Conflicts:
#	apps/dashboard/src/app/features/products/pages/product-list/products.page.ts
#	apps/dashboard/src/app/shared/ui/dynamic-form/components/dynamic-form.component.ts
#	apps/dashboard/src/app/shared/ui/dynamic-table/components/dynamic-table.component.html
#	apps/shell/public/assets/i18n/ar.json
#	apps/shell/public/assets/i18n/en.json

+++ apps/dashboard/src/app/features/products/pages/product-list/products.page.ts
@@ -1,4 +1,4 @@
-import { AfterViewInit, Component, inject, OnInit, signal, TemplateRef, ViewChild } from '@angular/core';
+import { AfterViewInit, Component, DestroyRef, inject, OnInit, signal, TemplateRef, ViewChild } from '@angular/core';
 import { ProductsService } from '../../services/products.service';
 import { ProductQueryParams } from '../../models/product-query.model';
 import { ProductModel } from '../../models/product.model';
@@ -10,10 +10,13 @@ import { TranslatePipe, TranslateService } from '@ngx-translate/core';
 import { ToastService } from '@org/shared-util-notification';
 import { ActivatedRoute, Router } from '@angular/router';
 import { ButtonComponent } from '@org/ui';
+import { DeleteConfirmationModalComponent } from '../../../../shared/ui/delete-confirmation-modal/delete-confirmation-modal.component';
+import { debounceTime, distinctUntilChanged, Subject, switchMap } from 'rxjs';
+import { takeUntilDestroyed } from '@angular/core/rxjs-interop';
 @Component({
   selector: 'app-products-page',
   standalone: true,
-  imports: [DynamicTableComponent ,LucideAngularModule ,TranslatePipe ,ButtonComponent],
+  imports: [DynamicTableComponent ,LucideAngularModule ,TranslatePipe ,ButtonComponent, DeleteConfirmationModalComponent],
   host: { class: 'flex flex-1 flex-col min-h-0' },
   templateUrl: './products.page.html',
   styleUrl: './products.page.scss',
@@ -26,30 +29,36 @@ export class ProductsPage implements OnInit , AfterViewInit{
   private readonly _translateService = inject(TranslateService);
   private readonly _router = inject(Router);
   private readonly _route = inject(ActivatedRoute);
+  private readonly _destroyRef = inject(DestroyRef);
+  
+  
+  private readonly _searchSubject = new Subject<string>();
 
   readonly Plus  = Plus ;
   products = signal<ProductModel[]>([]);
   page = signal(1);
   totalPages = signal(1);
+  readonly productToDelete = signal<ProductModel | null>(null);
 
   queryParams: ProductQueryParams = {
   page: 1,
   limit: 10,
   };
   columns: TableColumn<ProductModel>[] = [
-    {key: 'title',header: 'Name',},
-    {key: 'price',header: 'Price',},
-    {key: 'stock',header: 'Stock',},
-    {key: '_count', header: 'Sales',getValue: (row: ProductModel) => row._count.orderItems},
-    {key: 'ratings', header: 'Ratings',},
+    {key: 'title',header: 'DASHBOARD.PRODUCTS.TABLE.NAME',},
+    {key: 'price',header: 'DASHBOARD.PRODUCTS.TABLE.PRICE',},
+    {key: 'stock',header: 'DASHBOARD.PRODUCTS.TABLE.STOCK',},
+    {key: '_count', header: 'DASHBOARD.PRODUCTS.TABLE.SALES',getValue: (row: ProductModel) => row._count.orderItems},
+    {key: 'ratings', header: 'DASHBOARD.PRODUCTS.TABLE.RATINGS',},
     ];
 
   actions :TableAction<ProductModel>[] = [
-  { label: 'Edit',icon :Pencil, action: (product) => this.editProduct(product),},
-  { label: 'Delete',icon:Trash, action: (product) => this.deleteProduct(product),},
+  { label: 'DASHBOARD.PRODUCTS.ACTIONS.EDIT',icon :Pencil, action: (product) => this.editProduct(product),},
+  { label: 'DASHBOARD.PRODUCTS.ACTIONS.DELETE',icon:Trash, action: (product) => this.deleteProduct(product),},
   ];
    ngOnInit(): void {
     this.loadProducts();
+    this.setupSearch();
    }
 
   loadProducts(): void {
@@ -78,6 +87,16 @@ export class ProductsPage implements OnInit , AfterViewInit{
   }
 
   deleteProduct(product: ProductModel): void {
+    this.productToDelete.set(product);
+  }
+
+  confirmDelete(): void {
+    const product = this.productToDelete();
+
+    if (!product) {
+      return;
+    }
+
     this._productsService.deleteProduct(product.id).subscribe({
     next: () => {
       this._toastService.show(
@@ -85,18 +104,41 @@ export class ProductsPage implements OnInit , AfterViewInit{
       'success'
     );
       this.loadProducts();
+      this.productToDelete.set(null);
     },
   });
   }
 
-  onSearch(search: string): void {
-  this.queryParams = {
-    ...this.queryParams,
-    search,
-    page: 1,
-  };
+  closeDeleteModal(): void {
+    this.productToDelete.set(null);
+  }
+  private setupSearch(): void {
+  this._searchSubject
+    .pipe(
+      debounceTime(300),
+      distinctUntilChanged(),
+      switchMap((search) => {
+        this.queryParams = {
+          ...this.queryParams,
+          search,
+          page: 1,
+        };
 
-  this.loadProducts();
+        this.page.set(1);
+
+        return this._productsService.getProducts(this.queryParams);
+      }),
+      takeUntilDestroyed(this._destroyRef)
+    )
+    .subscribe({
+      next: (response) => {
+        this.products.set(response.payload.data);
+        this.totalPages.set(response.payload.metadata.totalPages);
+      },
+    });
+}
+  onSearch(search: string): void {
+    this._searchSubject.next(search);
    }
 
    onAddClicked(){
